refactor(search): split tuned-parameter store from optimizer, close last cycle
New src/core/search/tuning-store.ts holds the read/reset/apply side of
tuned parameters; tuning.ts keeps only the grid optimizer. search.ts
now depends on the store leaf, so search -> tuning -> benchmark ->
search no longer closes.
