+++ src/core/search/search.ts
@@ -82,7 +82,7 @@ import type {
   StructuredRecallQueryDocument,
 } from "./types.ts";
 import { expandQuery } from "./query-expansion.ts";
-import { applyTunedParameters, loadTunedParameters } from "./tuning.ts";
+import { applyTunedParameters, loadTunedParameters } from "./tuning-store.ts";
 import { resolveRecallProfile } from "./profiles.ts";
 import { applyCrossEncoderRerank } from "./rerank/index.ts";
 import { emitGatedTelemetry } from "../brain/continuity/emit.ts";
